# backend/docs/swagger.yaml
definitions:
...
  domain.LevelComplete:
    properties:
      average_velocity:
        type: number
      level_id:
        type: integer
      max_combo:
        type: integer
      num_press_err_by_char:
        additionalProperties:
          items:
            type: integer
          type: array
        description: NumPressErrByChar map[rune][2]int `json:"num_press_err_by_char"
          binding:"required" db:"-"`
        type: object
      placement:
        type: integer
      player_id:
        type: integer
      points:
        type: integer
    required:
    - average_velocity
    - level_id
    - max_combo
    - num_press_err_by_char
    - placement
    - player_id
    - points
    type: object
...
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for GoType game and website
  title: GoType App API
  version: 0.0.1
paths:
...
  /single-game/send-results:
    post:
      consumes:
      - application/json
      description: Send results of single player game
      operationId: send-results
      parameters:
      - description: Results of the game
        in: body
        name: complaint_id
        required: true
        schema:
          $ref: '#/definitions/domain.LevelComplete'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_NO_SUCH_USER - There is no user with
            such id'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;
            ERR_PERMISSION_DENIED - Not enough rights to perform the action'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Send results of single player game
      tags:
      - single-game
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"